Fix kconfighard integration
by probing for symlink already existing

Closes #1051

+++ classes/sca-kconfighard-recipe.bbclass
@@ -83,7 +83,8 @@ python do_sca_kconfighard() {
         tmp_result = os.path.join(d.getVar("T", True), "sca_raw_kconfighard.txt")
         d.setVar("SCA_RAW_RESULT_FILE", tmp_result)
 
-        os.symlink(os.path.join(d.getVar("B"), ".config"), os.path.join(d.getVar("B"), "config"))
+        if not os.path.exists(os.path.join(d.getVar("B"), "config")):
+            os.symlink(os.path.join(d.getVar("B"), ".config"), os.path.join(d.getVar("B"), "config"))
 
         _args = ["nativepython3", os.path.join(d.getVar("STAGING_BINDIR_NATIVE"), "kconfig-hardening-check", "kconfig-hardened-check.py")]      
         _args += ["-c", os.path.join(d.getVar("B"), ".config")]
